refactor library
This makes it possible to assemble directly into one struct with all necessary info for execution.

This PR also removes dependency on alloc from pio to make it even easier to use on embedded hardware. ArrayVec is used instead.

This PR is needed for rp-hal#74 and was written with the needs of rp-hal in mind.